requirements-win.txt
appdirs==1.4.3
cycler==0.10.0
et-xmlfile==1.0.1
h5py==2.7.0
jdcal==1.3
Keras==2.0.3
matplotlib==2.0.0
numpy==1.12.1+mkl
opencv-python==3.2.0
openpyxl==2.4.5
packaging==16.8
protobuf==3.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
sklearn==0.0
tensorflow-gpu==1.0.1
Theano==0.9.0
